requirements.txt
absl-py==1.3.0
Adafruit-Blinka==8.24.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-fingerprint==2.2.14
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.53.0
Adafruit-PureIO==1.1.11
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
APScheduler==3.10.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.0
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.2.0
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.13.0
django-meta==2.2.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.2.1
drf-yasg==1.21.4
EasyProcess==1.1
entrypoint2==1.1
executing==1.2.0
ez-setup==0.9
firebase-admin==6.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
flatbuffers==22.11.23
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
inflection==0.5.1
install==1.3.5
ipython==8.7.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
keyboard==0.13.5
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mongoengine==0.24.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.5
mss==9.0.1
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.13.1
openai==0.27.8
opencv-python==4.7.0.68
oscrypto==1.3.0
packaging==23.0
pandas==2.0.2
parso==0.8.3
pefile==2022.5.30
phonenumbers==8.13.4
pickleshare==0.7.5
Pillow==9.4.0
proglog==0.1.10
prompt-toolkit==3.0.33
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.21
pyftdi==0.55.0
PyGetWindow==0.0.9
Pygments==2.13.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
pymongo==4.3.3
PyMsgBox==1.0.9
pyparsing==3.0.9
pypdf==3.15.2
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
python-engineio==4.3.4
python-socketio==5.7.2
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.7
pytz==2022.7.1
pyusb==1.2.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzbar==0.1.8
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.3.0
social-auth-storage-mongoengine==1.0.1
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.4.0
svglib==1.5.1
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.65.0
traitlets==5.6.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
uuid==1.30
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.1.0
xhtml2pdf==0.2.11
yarl==1.9.2
